refactor: reuse useDataSource
