+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingLists/DataSourceSetup/EventWorkingListsDataSourceSetup.component.js
@@ -1,8 +1,6 @@
 // @flow
-import React, { useMemo } from 'react';
-import log from 'loglevel';
-import { errorCreator } from 'capture-core-utils';
-import { convertClientToList } from '../../../../../converters';
+import React from 'react';
+import { useDataSource } from '../../EventWorkingListsCommon';
 import { EventWorkingListsTemplateSetup } from '../TemplateSetup';
 import type { Props } from './eventWorkingListsDataSourceSetup.types';
 
@@ -11,57 +9,11 @@ export const EventWorkingListsDataSourceSetup = ({
     columns,
     recordsOrder,
     ...passOnProps
-}: Props) => {
-    const eventRecordsArray = useMemo(() =>
-        recordsOrder && eventRecords && recordsOrder
-            .map(eventId => ({
-                ...eventRecords[eventId],
-                eventId,
-            })), [
-        eventRecords,
-        recordsOrder,
-    ]);
-
-    const dataSource = useMemo(() => eventRecordsArray && eventRecordsArray
-        .map((eventRecord) => {
-            const listRecord = columns
-                .filter(column => column.visible)
-                .reduce((acc, { id, options, type }) => {
-                    const clientValue = eventRecord[id];
-
-                    if (options) {
-                        // TODO: Need is equal comparer for types because `sourceValue` and `option` can be an object for example (for some data element types) and we can't do strict comparison.
-                        const option = options.find(o => o.value === clientValue);
-                        if (!option) {
-                            log.error(
-                                errorCreator(
-                                    'Missing value in options')(
-                                    { id, clientValue, options }),
-                            );
-                        } else {
-                            acc[id] = option.text;
-                        }
-                    } else {
-                        acc[id] = convertClientToList(clientValue, type);
-                    }
-                    return acc;
-                }, {});
-
-            return {
-                ...listRecord,
-                eventId: eventRecord.eventId, // used as rowkey
-            };
-        }), [
-        eventRecordsArray,
-        columns,
-    ]);
-
-    return (
-        <EventWorkingListsTemplateSetup
-            {...passOnProps}
-            dataSource={dataSource}
-            columns={columns}
-            rowIdKey="eventId"
-        />
-    );
-};
+}: Props) => (
+    <EventWorkingListsTemplateSetup
+        {...passOnProps}
+        dataSource={useDataSource(eventRecords, recordsOrder, columns)}
+        columns={columns}
+        rowIdKey="eventId"
+    />
+);
