+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsCommon/index.js
@@ -1,6 +1,7 @@
 // @flow
 export { useDefaultColumnConfig } from './useDefaultColumnConfiguration';
 export { useColumns } from './useColumns';
+export { useDataSource } from './useDataSource';
 export type {
     EventRecords,
     MainColumnConfig,
