+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsCommon/useColumns/useColumns.js
@@ -3,7 +3,10 @@ import { useMemo } from 'react';
 import type { CustomColumnOrder } from '../../WorkingListsCommon';
 import type { EventWorkingListsColumnConfigs } from '../types';
 
-export const useColumns = (customColumnOrder?: CustomColumnOrder, defaultColumns: EventWorkingListsColumnConfigs): EventWorkingListsColumnConfigs => {
+export const useColumns = (
+    customColumnOrder?: CustomColumnOrder,
+    defaultColumns: EventWorkingListsColumnConfigs,
+): EventWorkingListsColumnConfigs => {
     const defaultColumnsAsObject = useMemo(() =>
         defaultColumns
             .reduce((acc, column) => ({ ...acc, [column.id]: column }), {}),
