+++ src/core_modules/capture-core/components/Pages/MainPage/EventWorkingListsCommon/useDataSource/index.js
@@ -1,0 +1,2 @@
+// @flow
+export { useDataSource } from './useDataSource';
